requirements.txt
torch>=1.10.0,<2.0.0
torchvision>=0.11.0,<0.15.0
numpy>=1.19.0
opencv-python>=4.5.0
scikit-learn>=0.24.0
scikit-image>=0.18.0
matplotlib>=3.4.0
pillow>=8.0.0
gradio>=3.0.0
tqdm>=4.60.0